refactor(ui): move geodata components into per-component folders
Aligns the geodata plugin with the component-per-folder convention in
.ai/rules/frontend.md and the layout used by sibling packages
(redisearch). Each component now lives in its own ComponentName/
directory with a barrel index.ts; consumers import from the central
src/components barrel. Pure rename + import-path refactor; no behavior
or test logic changes.
