+++ redisinsight/ui/src/packages/geodata/src/components/GeoHeader/index.ts
@@ -1,0 +1,1 @@
+export { GeoHeader } from './GeoHeader'
