+++ redisinsight/ui/src/packages/geodata/src/components/GeoPlot/index.ts
@@ -1,0 +1,1 @@
+export { GeoPlot } from './GeoPlot'
