+++ redisinsight/ui/src/packages/geodata/src/components/GeoTable/index.ts
@@ -1,0 +1,1 @@
+export { GeoTable } from './GeoTable'
