+++ redisinsight/ui/src/packages/geodata/src/components/Message/index.ts
@@ -1,0 +1,1 @@
+export { Message } from './Message'
