+++ redisinsight/ui/src/packages/geodata/src/components/RqeGeoVisualization/index.ts
@@ -1,0 +1,1 @@
+export { RqeGeoVisualization } from './RqeGeoVisualization'
